requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.1 